Snackbar added to Login/Signup/Recover containers.

+++ app/containers/Workspace/Workspace.js
@@ -27,7 +27,6 @@ class Workspace extends Component {
     this.toggleSettingsModal = this.toggleSettingsModal.bind(this);
     this.openFile = this.openFile.bind(this);
     this.addFile = this.addFile.bind(this);
-    this.deleteFile = this.deleteFile.bind(this);
     this.selectTab = this.selectTab.bind(this);
     this.closeTab = this.closeTab.bind(this);
     this.loadCodeSharing = this.loadCodeSharing.bind(this);
@@ -66,7 +65,9 @@ class Workspace extends Component {
   }
   deleteFile(data) {
     console.log('add file called with:', data);
-    this.props.deleteFile({project: {name: this.props.projectName}, data});
+    if(!data.path){
+      this.props.deleteFile({project: {name: this.props.projectName}, data});
+    }
   }
   openFile(file){
     let tabData = {
@@ -180,7 +181,7 @@ class Workspace extends Component {
           onSettingsClick={ this.toggleSettingsModal.bind(this, 'settingsOpen')  }
           addFile={ this.addFile }
           onFilesDrop={ this.onFilesDrop }
-          onFileDelete={ this.props.deleteFile }
+          onFileDelete={ this.deleteFile }
         />
         <Pane
           tabs={ this.props.tabs }
